+++ src/Content/Petabridge.Library/build.ps1
@@ -36,7 +36,7 @@ $DotNetInstallerUri = "https://raw.githubusercontent.com/dotnet/cli/v$DotNetVers
 $NugetVersion = "4.1.0";
 $NugetUrl = "https://dist.nuget.org/win-x86-commandline/v$NugetVersion/nuget.exe"
 $ProtobufVersion = "3.4.0"
-$DocfxVersion = "2.36.2"
+$DocfxVersion = "2.40.5"
 
 # Make sure tools folder exists
 $PSScriptRoot = Split-Path $MyInvocation.MyCommand.Path -Parent